add AbstractLayer class and simplify DataFrameTreeReduction
